requirements.txt
tqdm
wandb
hydra-core
numpy==1.24.3 # Required for compatibility with torch_geometric
torch
torch_geometric
toponetx
ogb 